Update README.md
The Driver Drowsiness Detection project aims to develop a system that monitors a driver’s level of alertness and detects signs of drowsiness. The system uses computer vision and deep learning techniques to analyze real-time video from a camera focused on the driver’s face, particularly the eyes. If the system detects that the driver’s eyes are closed for an extended period, indicating drowsiness, it triggers an alarm to alert the driver, thereby preventing potential accidents.